Every entry on one line in cards.cv.
Reads every line sequentially and uses these as input for the drinking
game. Uses an delimiter on the token list that is read from the file
cards.csv. Please use 25 or more entries even though only 24 is actually
used in the final pdf (this is because an error with the randomization).
Entries used in cards.csv are not to be taken seiously - was for testing
purposes only.